+++ lib/ibm/cloud/sdk/power_iaas.rb
@@ -1,7 +1,11 @@
+require_relative 'logging'
+
 module IBM
   module Cloud
     module SDK
       class PowerIaas < BaseService
+        include Logging
+
         # Create an API Client object for the Power IaaS service
         #
         # @param region [String] the IBM Power Cloud instance region
@@ -15,6 +19,7 @@ module IBM
           @region = region
           @token  = token
           @tenant = tenant
+          RestClient.log = logger
         end
 
         # Get the API service Endpoint URL